:sparkles: Support redirecting to other applications from the process
At times, a user task means that the user has to do something in a
different application. We support this with the form key 'zac:doRedirect',
and require a process variable 'redirectTo' to be set. If the variable
has a different name, input/output mapping in Camunda can be used on the
user task.

It is assumed that the other application requires some work to be done,
outside of the process modelling (highly specialized). For that,
whenever we prepare a redirect, we prepare an expected callback, and ask
to be redirect back to the Zaak detail page. The callback URL and return
URL are passed in the redirect as GET parameters. The callback URL is
meant to be POSTed back to, with the process variables that need to be
set. If these do not match the expected subsequent tasks, then the
input/output mapping can be used again.

TODO:

* currently, the callback URL is first-come, first serve. We use UUID to
  make it hard to guess, but it's not impossible. Stricter security
  checks (by signing the URL for example, to be validated by the client)
  should be done

* the API endpoint is currently not auth-protected, because we don't
  want to manage credentials. we rely on the one-time-token principle
  for the callback URL. To prevent abuse, we need to set up request
  throttling though (which protects against Denial-of-service attacks).

* the settings.DEBUG conditionals need to be removed, of course