:bug: Fix redirect flow for external-open tasks
The task completion is submitted in the ZAC, which completes the user task
and continues the process execution. However, the redirect URL was sending us
back to the Redirect task which would complete it as well, IF it didn't 404
because the user task did not exist anymore. The redirect URL that's
passed along now simply redirects to the zaak-detail page.