:sparkles: Implement closing activities
On closing of activities, they are automatically moved to the right
tab with finished activities. Action buttons are not available for
closed (not ongoing) activities.

Additionally, the component has been refactored for data-refreshing.
There is some optimization feasible (i.e. not loading the entire
list but rather the affected object, on changes).