:alien: Ensure that unknonwn URL-references don't crash ZAC detail page
Raising a `NotImplementedError` is great to be aware of issues, but it
makes the detail page effectively unusable if someone decides to mess
with the zaak betrokkenen data.

The behaviour changed to being more relaxed and displaying the problem
in the UI rather than crashing hard. Additionally, the table layout has
been fixed now to prevent overflows.