:lock: Moved Alfresco auth into component
For now, we depend on a username + password login for Alfresco. Initial
testing used an admin service account with credentials, but the
effect is that these eventually ended up in the DOM. Meaning, a user
without admin rights would very easily obtain the admin account
credentials.

Moving the username + password into the doc-selection modal fixes this
at the expense of having to do an 'extra' login. We do not store the
credentials for this user anywhere. This is how it's gonna be until
a better login system based on SSO machinery is implemented.