:lock: Filter zaken list based on OO restriction
This is a complex beast, requiring cache disabling because the cache
key becomes incredibly complex.

If OO restrictions are in place, then the initial candidates of zaken
are filtered on having a rol of betrokkene_type "organisatorische_eenheid".

This set of results is then filtered client-side based on the actual
related rol, where we retrieve if the rol of the specific OO is present
on each zaak in the result set.