:bug: Fix fetching zaken of wrong zaaktype
Permissions can be given to a zaaktype that do not necessarily allow
you to view zaken. The relevant permission filtering did not stop
the list query from happening. This did not falsely give access
to data that shouldn't have been visible, but it displayed the
records in the list while they shouldn't be displayed (because the
user doesn't have permission to view these zaken anyway).