:recycle: Refactor zaak suggestion autocomplete
It's now an actual auto-complete since partial identifications work.

Elastic Search index is used for the search/autocomplete functionality,
and it also contains the (immutable) properties identificatie,
bronorganisatie and URL to display in the frontend, which means we
have no need to hit the Zaken API at all for this functionality.