:bug: Fixed concurrency bug
When Kownsl advices/approvals are recorded, notifications are sent and
handled by the ZAC. These handlers comlete the user task in the
Camunda process. Sometimes, these notifications can arrive before the
redirect is triggered, which also completes the user task.

The code is now adapted for this race condition where the notification
is handled before the redirect. In that case, the return URL validates
against the historic process instance/task, and if it's a historic
entry, it no longer attempts to complete the user task (since that
has been done already)