:recycle: Composition over inheritance
Inheriting from a serializer with it's own specialized behaviour is
likely to have unintended side effects in the future. Especially
with a serializer that may be removed soon-ish.

Refactored base serializer into a custom serializer field with it's
own validator class. The validator only needs to get the relevant
zaak from the serializer context to be able to perform the validation.