:card_file_box: Squashed account migrations
This squashes all but the initial migration of the accounts app.

All RunPython operations have been removed, as they deal with legacy
data. Any new installations shall use the squashed migration file
and has an empty database to start with anyway, so the operations
are not relevant. Any existing installation shall use the original
migration files.